Check additional Uniswap pools for token prices
With the new sqrt price functionality in web30 we are able to check
Uniswap v3 swap prices in uncommon pool types while asserting that
there is sufficient liquidity to perform a swap between two tokens.

The new functionality will automatically generate an acceptable
sqrt price limit for all of the suggested Uniswap pools with
fees 0.01% 0.05% 0.3% and 1%, enabling profitable relaying of
newer tokens like PSTAKE with only a 1% pool to be detected. However,
if that 1% pool has very little liquidity, the price will be rejected.